pyproject.toml
[project]
name = "learnd"
version = "0.1.0"
description = "Adaptive Continuous Learning MCP Server with hierarchical vector storage using Qdrant Cloud and Mistral"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.3.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "mistralai[agents]>=1.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0"
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

# No scripts needed - use: uv run fastmcp run learnd.mcp_server:mcp

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]